chores: update eigen to 3.4.0 (#20)
* docs: typo + lint

* fix: force int for array indices #19

* fix: save block call in a vector before sample #19

* ci: workflow with eigen 3.4

* tests: add tests for multinomial

* ci: adjust tests parameter

* fix: initialize matrix to avoid unexpected values

* fix: initialize matrix

* typo: correct some typo

+++ MixtComp/src/lib/Mixture/Functional/FuncCSComputation.cpp
@@ -9,7 +9,7 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>
  **/
@@ -310,7 +310,7 @@ void globalQuantile(const Vector<FunctionCS>& vecInd, Vector<Real>& quantile) {
 	quantile(nQuantile - 1) = globalT(globalNTime - 1);
 
 	for (Index q = 1; q < nQuantile - 1; ++q) {
-		quantile(q) = globalT(q * quantileSize * (globalNTime - 1));
+		quantile(q) = globalT(int(q * quantileSize * (globalNTime - 1)));
 	}
 }
